test: relocate the 2 deferred Apple runner/recording tests (closes #980)
Completes the #983 relocation. The 2 files deferred there compute runtime fs
paths (fileURLToPath/__dirname), so they needed path-depth fixes on top of import
re-relativization:

- runner-client.test.ts -> src/platforms/apple/core/__tests__/ (joins the rest of
  the Apple engine tests); repoRoot recomputed one level deeper.
- recording-scripts.test.ts -> src/recording/__tests__/ (beside the overlay.ts it
  tests); ios-runner RecordingScripts + test/integration/support paths recomputed
  one level shallower.

src/platforms/ios/__tests__/ is now empty (all Apple-engine tests live beside
their source under apple/core and recording). Pure test relocation — tsc, oxlint,
oxfmt, layering guard, fallow green; full unit suite passes (2883).
